+++ backend/src/modules/auth/auth.service.ts
@@ -54,6 +54,10 @@ export class AuthService {
       throw new ForbiddenException('Account is deactivated');
     }
 
+    if (!user.emailVerified) {
+      throw new ForbiddenException('Email not verified');
+    }
+
     // If 2FA is enabled, issue a short-lived pre-auth token instead of a full access token
     if (user.twoFactorEnabled) {
       const preAuthToken = this.jwtService.sign(
